feat(auth): wire conditional Google OAuth provider + login button
Les credentials GOOGLE_CLIENT_ID/GOOGLE_CLIENT_SECRET fournis pour la démo
étaient valides mais non câblés : NextAuth n'exposait qu'un provider
Credentials. Ce commit déclare le provider Google **uniquement** quand les
deux secrets sont présents en env (parité « ship-without-keys » — l'app boote
identiquement sans clé), et affiche un bouton « Continuer avec Google » sur la
page de login dans ce seul cas.

- src/lib/auth/config.ts : provider Google conditionnel (spread sur googleEnabled),
  allowDangerousEmailAccountLinking activé pour la démo (lie Google à un compte
  email/mdp existant de même email ; à durcir avant prod pure).
- src/app/(auth)/login/page.tsx : server component, lit l'env et passe
  googleEnabled au formulaire.
- src/app/(auth)/login/login-form.tsx : extraction client du formulaire +
  bouton Google (signIn("google")).

Prérequis côté Google Cloud Console (action opérateur) : enregistrer le redirect
URI autorisé https://<URL>/api/auth/callback/google.
