+++ src/app/(auth)/login/page.tsx
@@ -1,188 +1,22 @@
-"use client";
-
-import { Suspense, useState } from "react";
-import { signIn, getSession } from "next-auth/react";
-import { useSearchParams, useRouter } from "next/navigation";
-import Link from "next/link";
-
-function portalForRole(role: string | undefined): string {
-  switch (role) {
-    case "ADMIN":
-    case "OPERATOR":
-      return "/console";
-    case "FOUNDER":
-    case "BRAND":
-    case "CLIENT_RETAINER":
-    case "CLIENT_STATIC":
-      return "/cockpit";
-    case "AGENCY":
-      return "/agency";
-    case "CREATOR":
-    case "FREELANCE":
-      return "/creator";
-    case "USER":
-    default:
-      // Nouveau compte ou role non spécialisé → hub des portails accessibles
-      // (cockpit + creator par défaut). Cf. src/proxy.ts COCKPIT_ROLES/CREATOR_ROLES.
-      return "/portals";
-  }
-}
-
-function LoginForm() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const callbackUrl = searchParams.get("callbackUrl");
-
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState<string | null>(null);
-  const [loading, setLoading] = useState(false);
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    setError(null);
-    setLoading(true);
-
-    try {
-      const result = await signIn("credentials", {
-        email,
-        password,
-        redirect: false,
-      });
-
-      if (result?.error) {
-        setError("Email ou mot de passe invalide.");
-        setLoading(false);
-        return;
-      }
-
-      // Si callbackUrl fourni (deep link), on respecte. Sinon, redirect
-      // role-based vers le portail correspondant.
-      let target = callbackUrl;
-      if (!target) {
-        const session = await getSession();
-        target = portalForRole(session?.user?.role);
-      }
-      router.push(target);
-      router.refresh();
-    } catch {
-      setError("Une erreur est survenue. Veuillez reessayer.");
-      setLoading(false);
-    }
-  }
-
-  return (
-    <div className="flex flex-col items-center">
-      {/* Branding */}
-      <div className="mb-8 text-center">
-        <div className="mb-3 flex items-center justify-center gap-2">
-          <div className="flex h-10 w-10 items-center justify-center rounded-xl bg-violet-500/20">
-            <svg
-              className="h-5 w-5 text-violet-400"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-              strokeWidth={2}
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M13 10V3L4 14h7v7l9-11h-7z"
-              />
-            </svg>
-          </div>
-          <h1 className="text-2xl font-bold text-white">LaFusee</h1>
-        </div>
-        <p className="text-sm text-zinc-500">
-          Industry OS — De la Poussiere a l&apos;Etoile
-        </p>
-      </div>
-
-      {/* Login Form */}
-      <div className="w-full rounded-xl border border-zinc-800 bg-zinc-900/50 p-6 shadow-xl">
-        <h2 className="mb-6 text-center text-lg font-semibold text-white">
-          Connexion
-        </h2>
-
-        <form onSubmit={handleSubmit} className="flex flex-col gap-4">
-          {error && (
-            <div className="rounded-lg border border-red-800/50 bg-red-950/50 px-4 py-3 text-sm text-red-300">
-              {error}
-            </div>
-          )}
-
-          <div>
-            <label
-              htmlFor="email"
-              className="mb-1.5 block text-sm font-medium text-zinc-300"
-            >
-              Email
-            </label>
-            <input
-              id="email"
-              name="email"
-              type="email"
-              autoComplete="email"
-              required
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              placeholder="vous@entreprise.com"
-              className="w-full rounded-lg border border-zinc-700 bg-zinc-800 px-3 py-2.5 text-sm text-white placeholder:text-zinc-500 focus:border-violet-500 focus:outline-none focus:ring-1 focus:ring-violet-500"
-            />
-          </div>
-
-          <div>
-            <div className="mb-1.5 flex items-center justify-between">
-              <label
-                htmlFor="password"
-                className="block text-sm font-medium text-zinc-300"
-              >
-                Mot de passe
-              </label>
-              <Link href="/forgot-password" className="text-xs text-violet-400 hover:text-violet-300">
-                Mot de passe oublie ?
-              </Link>
-            </div>
-            <input
-              id="password"
-              name="password"
-              type="password"
-              autoComplete="current-password"
-              required
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              placeholder="••••••••"
-              className="w-full rounded-lg border border-zinc-700 bg-zinc-800 px-3 py-2.5 text-sm text-white placeholder:text-zinc-500 focus:border-violet-500 focus:outline-none focus:ring-1 focus:ring-violet-500"
-            />
-          </div>
-
-          <button
-            type="submit"
-            disabled={loading}
-            className="mt-2 w-full rounded-lg bg-violet-600 px-4 py-2.5 text-sm font-semibold text-white transition-colors hover:bg-violet-500 focus:outline-none focus:ring-2 focus:ring-violet-500 focus:ring-offset-2 focus:ring-offset-zinc-900 disabled:cursor-not-allowed disabled:opacity-50"
-          >
-            {loading ? "Connexion en cours..." : "Se connecter"}
-          </button>
-        </form>
-      </div>
-
-      {/* Register link */}
-      <div className="mt-6 text-center">
-        <p className="text-sm text-zinc-400">
-          Pas encore de compte ?{" "}
-          <Link href="/register" className="font-medium text-violet-400 hover:text-violet-300">
-            Creer un compte
-          </Link>
-        </p>
-      </div>
-    </div>
-  );
-}
+import { Suspense } from "react";
+import { LoginForm } from "./login-form";
 
 export default function LoginPage() {
+  // Lu côté serveur : le bouton Google n'apparaît que si les deux secrets sont
+  // présents en env (parité « ship-without-keys », cf. src/lib/auth/config.ts).
+  const googleEnabled = Boolean(
+    process.env.GOOGLE_CLIENT_ID && process.env.GOOGLE_CLIENT_SECRET,
+  );
+
   return (
-    <Suspense fallback={<div className="flex items-center justify-center text-zinc-500">Chargement...</div>}>
-      <LoginForm />
+    <Suspense
+      fallback={
+        <div className="flex items-center justify-center text-zinc-500">
+          Chargement...
+        </div>
+      }
+    >
+      <LoginForm googleEnabled={googleEnabled} />
     </Suspense>
   );
 }
